Make sure to run Apollo set up code on run time.
